moved _gen_even_slices to utils/init add test codes for checking callable init
